requirements.txt
flask==2.3.3
googlemaps==4.10.0
gspread==5.12.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
requests==2.31.0
python-dotenv==1.0.0
google-auth[service-account]==2.23.4


